refactor: decompose ws-server helper clusters into src/lib/ws-server/
Extract cohesive, low-coupling helper clusters from the ws-server
monolith into modules under src/lib/ws-server/ (faithful moves, no
behavior change; all well under the 800-line ceiling):

  - pty-support.ts  — sanitizePtyEnv, resolvePreferredShell,
    resolveTmuxBinary, tmuxSessionExists, isDashTerminalSession
    (the seam the terminal-host extraction shares)
  - channels.ts     — BACKPRESSURE_* constants + isLossyMessage
    (the load-bearing lossy/durable classification)
  - next-fetch.ts   — Next API fetch/retry/backoff + JSON envelope
    (self-sources the ws token via cached getOrCreateWsToken)
  - git-worktrees.ts — GitWorktreeRecord, shortHome, parseGitWorktreeList

The entry file drops ~170 lines and stays the multiplexer (which
CLAUDE.md waives from the ceiling); the event-loop singleton core
(clients/send/broadcast/wss/realtime) is deliberately left in place to
avoid regression risk in the freeze-sensitive path. Verified: tsc clean,
eslint clean, 1344 tests green, ws-server boots on :3999 with /health,
and the tauri-export esbuild step bundles the decomposed entry
(node --check passes on the 2.7mb ws-server.mjs).
